+++ common/src/main/kotlin/com/cobblemon/mod/common/client/render/models/blockbench/pokemon/gen3/VolbeatModel.kt
@@ -8,11 +8,13 @@
 
 package com.cobblemon.mod.common.client.render.models.blockbench.pokemon.gen3
 
+import com.cobblemon.mod.common.client.render.models.blockbench.createTransformation
 import com.cobblemon.mod.common.client.render.models.blockbench.frame.BiWingedFrame
 import com.cobblemon.mod.common.client.render.models.blockbench.frame.BipedFrame
 import com.cobblemon.mod.common.client.render.models.blockbench.frame.HeadedFrame
 import com.cobblemon.mod.common.client.render.models.blockbench.pokemon.PokemonPose
 import com.cobblemon.mod.common.client.render.models.blockbench.pokemon.PokemonPoseableModel
+import com.cobblemon.mod.common.client.render.models.blockbench.pose.ModelPartTransformation
 import com.cobblemon.mod.common.entity.PoseType
 import net.minecraft.client.model.ModelPart
 import net.minecraft.util.math.Vec3d
@@ -35,34 +37,57 @@ class VolbeatModel (root: ModelPart) : PokemonPoseableModel(), HeadedFrame, Bipe
     lateinit var stand: PokemonPose
     lateinit var walk: PokemonPose
     lateinit var hover: PokemonPose
+    lateinit var water_surface_idle: PokemonPose
+    lateinit var water_surface_fly: PokemonPose
     lateinit var fly: PokemonPose
     lateinit var battleidle: PokemonPose
 
+    val wateroffset = -15
+
     override fun registerPoses() {
         val blink = quirk("blink") { bedrockStateful("volbeat", "blink").setPreventsIdle(false) }
+        val flicker = quirk("flicker") { bedrockStateful("volbeat", "flicker_quirk").setPreventsIdle(false) }
 
         sleep = registerPose(
             poseType = PoseType.SLEEP,
             idleAnimations = arrayOf(bedrock("volbeat", "sleep"))
         )
 
-        stand = registerPose(
-            poseName = "standing",
-            poseTypes = PoseType.UI_POSES + PoseType.STAND,
+        water_surface_idle = registerPose(
+            poseName = "water_surface",
+            poseTypes = PoseType.STATIONARY_POSES,
             transformTicks = 10,
-            quirks = arrayOf(blink),
-            condition = { !it.isBattling && !it.isTouchingWater},
+            quirks = arrayOf(blink, flicker),
+            condition = { it.isTouchingWater && !it.isSubmergedInWater },
+            transformedParts = arrayOf(
+                rootPart.createTransformation().addPosition(ModelPartTransformation.Y_AXIS, wateroffset)
+            ),
             idleAnimations = arrayOf(
                 singleBoneLook(),
-                bedrock("volbeat", "ground_idle")
+                bedrock("volbeat", "air_idle")
+            )
+        )
+
+        water_surface_fly = registerPose(
+            poseName = "water_surface_fly",
+            poseTypes = PoseType.MOVING_POSES,
+            transformTicks = 10,
+            quirks = arrayOf(blink, flicker),
+            condition = { it.isTouchingWater && !it.isSubmergedInWater },
+            transformedParts = arrayOf(
+                rootPart.createTransformation().addPosition(ModelPartTransformation.Y_AXIS, wateroffset)
+            ),
+            idleAnimations = arrayOf(
+                singleBoneLook(),
+                bedrock("volbeat", "air_fly")
             )
         )
 
         hover = registerPose(
             poseName = "hover",
-            poseTypes = setOf(PoseType.HOVER, PoseType.FLOAT),
+            poseTypes = setOf(PoseType.HOVER),
             transformTicks = 10,
-            quirks = arrayOf(blink),
+            quirks = arrayOf(blink, flicker),
             idleAnimations = arrayOf(
                 singleBoneLook(),
                 bedrock("volbeat", "air_idle")
@@ -71,12 +96,24 @@ class VolbeatModel (root: ModelPart) : PokemonPoseableModel(), HeadedFrame, Bipe
 
         fly = registerPose(
             poseName = "fly",
-            poseTypes = setOf(PoseType.FLY, PoseType.SWIM),
+            poseTypes = setOf(PoseType.FLY),
             transformTicks = 10,
-            quirks = arrayOf(blink),
+            quirks = arrayOf(blink, flicker),
             idleAnimations = arrayOf(
                 singleBoneLook(),
                 bedrock("volbeat", "air_fly")
+            )
+        )
+
+        stand = registerPose(
+            poseName = "standing",
+            poseTypes = PoseType.UI_POSES + PoseType.STATIONARY_POSES,
+            transformTicks = 10,
+            quirks = arrayOf(blink, flicker),
+            condition = { !it.isBattling},
+            idleAnimations = arrayOf(
+                singleBoneLook(),
+                bedrock("volbeat", "ground_idle")
             )
         )
 
@@ -84,8 +121,7 @@ class VolbeatModel (root: ModelPart) : PokemonPoseableModel(), HeadedFrame, Bipe
             poseName = "walking",
             poseType = PoseType.WALK,
             transformTicks = 10,
-            quirks = arrayOf(blink),
-            condition = { !it.isTouchingWater },
+            quirks = arrayOf(blink, flicker),
             idleAnimations = arrayOf(
                 singleBoneLook(),
                 bedrock("volbeat", "ground_walk")
@@ -96,7 +132,7 @@ class VolbeatModel (root: ModelPart) : PokemonPoseableModel(), HeadedFrame, Bipe
             poseName = "battle_idle",
             poseTypes = PoseType.STATIONARY_POSES,
             transformTicks = 10,
-            quirks = arrayOf(blink),
+            quirks = arrayOf(blink, flicker),
             condition = { it.isBattling },
             idleAnimations = arrayOf(
                 singleBoneLook(),